+++ test/external/modules/view-licences/helpers.js
@@ -1,176 +1,11 @@
 'use strict';
 
 const Lab = require('lab');
-const { experiment, test, beforeEach, afterEach } = exports.lab = Lab.script();
+const { experiment, test } = exports.lab = Lab.script();
 
 const { expect } = require('code');
-const sinon = require('sinon');
 
-const { selectRiverLevelMeasure, loadRiverLevelData, mapFilter, getLicencePageTitle, setConditionHofFlags } = require('../../../../src/external/modules/view-licences/helpers');
-const services = require('external/lib/connectors/services');
-
-const getTestMeasure = (parameter = 'flow', hasLatestReading = true) => {
-  const latestReading = hasLatestReading
-    ? { value: 0.321 }
-    : 'http://example.com';
-
-  return {
-    id: `test-${parameter}`,
-    parameter,
-    latestReading
-  };
-};
-
-const getLicenceData = (code1, subCode1, code2, subCode2) => {
-  return { pageTitle: 'test',
-    summary: {
-      conditions: [{
-        code: code1,
-        subCode: subCode1,
-        otherData: 'should remain',
-        test: 1234
-      }, {
-        code: code2,
-        subCode: subCode2,
-        otherData: 'should also remain',
-        test: 9876
-      }]
-    } };
-};
-
-experiment('selectRiverLevelMeasure', () => {
-  test('returns the first measure if there is only one', async () => {
-    const riverLevel = {
-      measures: [getTestMeasure()]
-    };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure.id).to.equal('test-flow');
-  });
-
-  test('returns undefined if the measure does not contain a latest reading', async () => {
-    const riverLevel = { measures: [getTestMeasure('flow', false)] };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure).to.be.undefined();
-  });
-
-  test('returns level if cesLev is true and cesFlow is false', async () => {
-    const riverLevel = {
-      measures: [
-        getTestMeasure('flow'),
-        getTestMeasure('level')
-      ]
-    };
-    const hofTypes = { cesFlow: false, cesLev: true };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure.id).to.equal('test-level');
-  });
-
-  test('returns undefined if cesLev is true and cesFlow is false, but no latest reading', async () => {
-    const riverLevel = {
-      measures: [
-        getTestMeasure('flow', false),
-        getTestMeasure('level', false)
-      ]
-    };
-    const hofTypes = { cesFlow: false, cesLev: true };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure).to.be.undefined();
-  });
-
-  test('returns flow if cesLev is false and cesFlow is true', async () => {
-    const riverLevel = {
-      measures: [
-        getTestMeasure('flow'),
-        getTestMeasure('level')
-      ]
-    };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure.id).to.equal('test-flow');
-  });
-
-  test('returns undefined if cesLev is false and cesFlow is true, but no latest reading', async () => {
-    const riverLevel = {
-      measures: [
-        getTestMeasure('flow', false),
-        getTestMeasure('level', false)
-      ]
-    };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    const measure = selectRiverLevelMeasure(riverLevel, hofTypes);
-    expect(measure).to.be.undefined();
-  });
-});
-
-experiment('loadRiverLevelData', () => {
-  const hofTypes = { cesLevel: true, cesFlow: false };
-
-  beforeEach(async () => {
-    sinon.stub(services.water.riverLevels, 'getRiverLevel');
-  });
-
-  afterEach(async () => {
-    services.water.riverLevels.getRiverLevel.restore();
-  });
-
-  test('returns null riverLevel and measure when no station reference', async () => {
-    const riverLevelData = await loadRiverLevelData(null, hofTypes);
-    expect(riverLevelData).to.equal({ riverLevel: null, measure: null });
-  });
-
-  test('returns the river level and measure when the station is returned', async () => {
-    const response = {
-      measures: [{
-        latestReading: { value: 21.7 },
-        parameter: 'flow',
-        valueType: 'instantaneous'
-      }]
-    };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    services.water.riverLevels.getRiverLevel.resolves(response);
-    const riverLevelData = await loadRiverLevelData(1234, hofTypes);
-    expect(riverLevelData.measure).to.be.an.object();
-    expect(riverLevelData.riverLevel).to.be.an.object();
-  });
-
-  test('does not return flow if only level HoF type in licence', async () => {
-    const response = {
-      measures: [{
-        latestReading: { value: 21.7 },
-        parameter: 'flow',
-        valueType: 'instantaneous'
-      }]
-    };
-    const hofTypes = { cesFlow: false, cesLev: true };
-    services.water.riverLevels.getRiverLevel.resolves(response);
-    const riverLevelData = await loadRiverLevelData(1234, hofTypes);
-    expect(riverLevelData.measure).to.be.undefined();
-    expect(riverLevelData.riverLevel).to.be.an.object();
-  });
-
-  test('does not return level if only flow HoF type in licence', async () => {
-    const response = {
-      measures: [{
-        latestReading: { value: 21.7 },
-        parameter: 'level',
-        valueType: 'instantaneous'
-      }]
-    };
-    const hofTypes = { cesFlow: true, cesLev: false };
-    services.water.riverLevels.getRiverLevel.resolves(response);
-    const riverLevelData = await loadRiverLevelData(1234, hofTypes);
-    expect(riverLevelData.measure).to.be.undefined();
-    expect(riverLevelData.riverLevel).to.be.an.object();
-  });
-
-  test('returns null riverLevel and measure when no station is found', async () => {
-    services.water.riverLevels.getRiverLevel.rejects({ statusCode: 404 });
-    const riverLevelData = await loadRiverLevelData(null, hofTypes);
-    expect(riverLevelData).to.equal({ riverLevel: null, measure: null });
-  });
-});
+const { mapFilter } = require('../../../../src/external/modules/view-licences/helpers');
 
 experiment('mapFilter', () => {
   test('adds the company entity id to the filter', async () => {
@@ -208,74 +43,5 @@ experiment('mapFilter', () => {
 
     filter = mapFilter('1234', { emailAddress: 'right@example.com  ' });
     expect(filter.email).to.equal('right@example.com');
-  });
-});
-
-experiment('getLicencePageTitle', () => {
-  test('if licence has custom name, return "Licence number 1234"', async () => {
-    const { pageTitle, pageHeading } = getLicencePageTitle('nunjucks/view-licences/contact.njk', '1234', 'customName');
-    expect(pageHeading).to.equal('Licence number 1234');
-    expect(pageTitle).to.equal('Contact details for customName');
-  });
-
-  test('if licence does not have custom name, return title for key', async () => {
-    const { pageTitle, pageHeading } = getLicencePageTitle('nunjucks/view-licences/points.njk', '1234');
-    expect(pageTitle).to.equal('Abstraction points for 1234');
-    expect(pageHeading).to.equal('Abstraction points for licence number 1234');
-  });
-
-  test('if licence has custom name and unexpected key, return default response', async () => {
-    const { pageTitle, pageHeading } = getLicencePageTitle('nunjucks/view-licences/test', '1234', 'customName');
-    expect(pageTitle).to.equal('Licence number 1234');
-    expect(pageHeading).to.equal('Licence number 1234');
-  });
-
-  test('if licence does not have custom name and unexpected key, return default response', async () => {
-    const { pageTitle, pageHeading } = getLicencePageTitle('nunjucks/view-licences/test', '1234');
-    expect(pageTitle).to.equal('Licence number 1234');
-    expect(pageHeading).to.equal('Licence number 1234');
-  });
-});
-
-experiment('setConditionHofFlags', () => {
-  test('creates isHof flag for each condition', async () => {
-    const result = setConditionHofFlags(getLicenceData('CES', 'LEV', 'CES', 'FLOW'));
-    expect(result.summary.conditions[0]).to.include('isHof');
-    expect(result.summary.conditions[1]).to.include('isHof');
-  });
-
-  test('retains existing condition data', async () => {
-    const existingConditionData = ['code', 'subCode', 'otherData', 'test'];
-    const result = setConditionHofFlags(getLicenceData('CES', 'LEV', 'CES', 'FLOW'));
-    expect(result.summary.conditions[0]).to.include(existingConditionData);
-    expect(result.summary.conditions[1]).to.include(existingConditionData);
-  });
-
-  experiment('isHof value', () => {
-    test('flags are set independently from each other', async () => {
-      const result = setConditionHofFlags(getLicenceData('OTHER', 'LEV', 'CES', 'FLOW'));
-      expect(result.summary.conditions[0].isHof).to.be.false();
-      expect(result.summary.conditions[1].isHof).to.be.true();
-    });
-
-    test('is set to false when code !== "CES"', async () => {
-      const result = setConditionHofFlags(getLicenceData('OTHER', 'LEV'));
-      expect(result.summary.conditions[0].isHof).to.be.false();
-    });
-
-    test('is set to false when subCode !== "FLOW" or "LEV"', async () => {
-      const result = setConditionHofFlags(getLicenceData('CES', 'OTHER'));
-      expect(result.summary.conditions[0].isHof).to.be.false();
-    });
-
-    test('is set to true when code !== "CES" subCode !== "FLOW"', async () => {
-      const result = setConditionHofFlags(getLicenceData('CES', 'FLOW'));
-      expect(result.summary.conditions[0].isHof).to.be.true();
-    });
-
-    test('is set to true when code !== "CES" subCode !== "LEV"', async () => {
-      const result = setConditionHofFlags(getLicenceData('CES', 'LEV'));
-      expect(result.summary.conditions[0].isHof).to.be.true();
-    });
   });
 });
