+++ test/shared/plugins/licence-data/index.js
@@ -15,8 +15,7 @@ experiment('Licence data plugin', () => {
       continue: 'continue',
       realm: {
         pluginOptions: {
-          getSummaryByDocumentId: sandbox.stub(),
-          getCommunicationsByDocumentId: sandbox.stub()
+          getLicenceData: sandbox.stub().resolves({ error: null })
         }
       }
     };
@@ -88,8 +87,10 @@ experiment('Licence data plugin', () => {
       test('loads data using getSummaryByDocumentId', async () => {
         await plugin._onPreHandler(request, h);
 
-        expect(h.realm.pluginOptions.getSummaryByDocumentId.calledWith(
-          request.params.documentId, request
+        expect(h.realm.pluginOptions.getLicenceData.calledWith(
+          'getSummaryByDocumentId',
+          request.params.documentId,
+          request
         )).to.equal(true);
       });
 
@@ -99,7 +100,7 @@ experiment('Licence data plugin', () => {
       });
 
       test('throws a Boom error and logs if getter throws error', async () => {
-        h.realm.pluginOptions.getSummaryByDocumentId.rejects();
+        h.realm.pluginOptions.getLicenceData.rejects();
 
         try {
           await plugin._onPreHandler(request, h);
@@ -128,8 +129,10 @@ experiment('Licence data plugin', () => {
       test('loads data using getCommunicationsByDocumentId', async () => {
         await plugin._onPreHandler(request, h);
 
-        expect(h.realm.pluginOptions.getCommunicationsByDocumentId.calledWith(
-          request.params.documentId, request
+        expect(h.realm.pluginOptions.getLicenceData.calledWith(
+          'getCommunicationsByDocumentId',
+          request.params.documentId,
+          request
         )).to.equal(true);
       });
     });
